style: GlobalStyle.js wrapper 분리
