+++ src/style/GlobalStyle.tsx
@@ -12,11 +12,15 @@ const GlobalStyle = createGlobalStyle`
         justify-content: center;
         align-items: center;
     }
-    .pageWrapper{
+    .mainWrapper{
         width: 23.4375rem;
         height: 50.75rem;
         border-radius: 1.25rem;
         background-color : var(--black);
+    }
+    .pageWrapper{ /*pageWrapper는 header footer 제외 높이*/
+        width: 23.4375rem;
+        height: 45.6875rem;
     }
 `;
